+++ server/go.mod
@@ -3,7 +3,8 @@ module github.com/rawdaGastan/grid3_auto_deployer
 go 1.16
 
 require (
-	github.com/mattn/go-sqlite3 v1.14.16
-	gorm.io/driver/sqlite v1.4.4 // indirect
+	github.com/mattn/go-sqlite3 v1.14.16 // indirect
+	github.com/threefoldtech/grid3-go v0.0.0-20230214163319-124637fb2909
+	gorm.io/driver/sqlite v1.4.4
 	gorm.io/gorm v1.24.5
 )
