+++ server/models/database.go
@@ -1,31 +1,19 @@
 package models
 
 import (
-	"fmt"
-
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 )
 
-type User struct {
-	ID       string `json:"id"`
-	Name     string `json:"name"`
-	Mail     string `json:"mail"`
-	Password string `json:"password"`
-	Voucher  string `json:"voucher"`
+type database struct {
+	db *gorm.DB
 }
 
-type Quota struct {
-	UserID string `json:"userID"`
-	Vms    int    `json:"vms"`
-	K8s    int    `json:"k8s"`
-}
-
-func ConnectDB() (*gorm.DB, error) {
-	db, err := gorm.Open(sqlite.Open("./database.db"), &gorm.Config{})
+func (d *database) ConnectDB(file string) (err error) {
+	d.db, err = gorm.Open(sqlite.Open(file), &gorm.Config{})
 	if err != nil {
-		fmt.Printf("Failed to Connect to database")
+		return err
 	}
-
-	return db, err
+	err = d.db.AutoMigrate(&User{}, &Quota{})
+	return err
 }
